requirements.txt
Django==2.1.7
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-extra-fields==1.1.0
django-filer==1.4.4
django-filter==2.1.0
django-js-asset==1.2.2
django-money==0.7.3
django-mptt==0.10.0
django-pg-current-timestamp==0.2.4
django-polymorphic==2.0.3
django-simple-history==1.9.1
django-smart-selects==1.5.4
django-treewidget==0.3.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
easy-thumbnails==2.6
Markdown==3.1.1
Pillow==6.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py-moneyed==0.8.0
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
sorl-thumbnail==12.5.0
Unidecode==1.0.23
